Make platformio.ini a project root marker file
Apparently, the file given when registering the project type in
projectile is only used to determine the type of project, it is not
used to find project root directories.

If this is not done, every platformio project will need to be in the
root of a git repository, or contain a .projectile file in the root directory.